backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
numpy==1.26.4
Pillow>=10.3.0
gunicorn==21.2.0
google-generativeai>=0.1.0
